feat(auth-0): replace LoginForm component instead Login component
